// apps/hojas_de_vida/src/app/core/models/persona.model.ts
export interface Persona {
  id?: string;
  primerNombre: string;
  segundoNombre?: string;
  primerApellido: string;
  segundoApellido?: string;
  tipoDocumento: string;
  identificacion: string;
  paisExpedicion: string;
  departamentoExpedicion?: string;
  ciudadExpedicion: string;
  fechaExpedicion: Date;
  fechaNacimiento: Date;
  paisNacimiento: string;
  departamentoNacimiento?: string;
  ciudadNacimiento: string;
  genero: string;
  estadoCivil: string;
  nacionalidad: string;
  etnia: string;
  discapacidad: string;
  victimaConflicto: boolean;
  paisResidencia: string;
  departamentoResidencia?: string;
  ciudadResidencia: string;
  barrio: string;
  direccion: string;
  sector: string;
  telefono?: string;
  celular1: string;
  celular2?: string;
  correo: string;
  correoPersonal?: string; 
  aspiracionSalario?: number;
  trabajoActual?: boolean;
  area: string;
  hojaVidaPresentada?: boolean;
  trabajaUniversidad?: boolean;
  correoInstitucional?: string;
  cargoUniversidad?: string;
  dependencia?: string;
  egresadoUniversidad?: boolean;
  filiacionPolitica?: string;
  estrato?: string;
  tieneLibretaMilitar?: boolean;
  libretaMilitar?: string;
  clase?: number;
  distrito?: string;
  politica_datos?: boolean;
  notificacionesConvocatoriasEmail?: boolean; 
  fechaPreferenciaNotificaciones?: Date; 
  informacionAcademica?: any[];
  informacionLaboral?: any[];
  informacionesFamiliares?: any[];
  competencias?: any[];
  archivosSubidos?: any[];
  afiliaciones?: any[];
  referencias?: any[];
  documentosSoporte?: any[];
  idAuthUsuario?: string;
  sinExperienciaLaboral?: boolean;
  fechaDeclaracionSinExperienciaLaboral?: string | null;
}